fix: fix typos in the song name (Helloween - tale that wasn't right), fix problem with Stone Sour Wicked game song album cover, fix typos in the self-check of the audio-player task;  refactor: split logic of the audio-player into separate modules with pure functions and imperative function wrappers
